correlation: move impl Event for Message closer to Message
